kv cache

+++ src/model/model_loader.py
@@ -41,7 +41,9 @@ def load_model_and_tokenizer(
         # Choose default torch_dtype if not provided
         if torch_dtype is None:
             if device_override in ("cuda", "mps"):
-                torch_dtype = torch.float16
+                #torch_dtype = torch.float16
+                #torch_dtype = torch.bfloat16
+                torch_dtype = torch.float32
             else:
                 torch_dtype = torch.float32
